services FTW! UserService now fully contained in UserActions
