+++ bc-react/src/actions/UserActions.js
@@ -1,5 +1,12 @@
 import dispatcher from '../Dispatcher'
 
+let baseUrl;
+if(process.env.NODE_ENV === 'production'){
+  baseUrl = "/"
+} else {
+  baseUrl = "http://localhost:4000/"
+}
+
 export function updateRegistration(attribute, value){
   dispatcher.dispatch({
     type: 'UPDATE_REGISTRATION',
@@ -28,12 +35,36 @@ export function submitLogin(loginAttributes){
   })
 }
 
-export function updateUser(attributes){
-  dispatcher.dispatch({
-    type: 'UPDATE_USER',
-    attributes: attributes
+export function processLogin(attributes){
+  const params = {
+    method: 'PUT',
+    headers: {'Content-Type': 'application/json'},
+    body: JSON.stringify(attributes)
+  }
+  fetch(`${baseUrl}login`, params).then((response)=>{
+    if(response.ok){
+      response.json().then((body)=>{
+        // updateUser(body.user)
+        dispatcher.dispatch({
+          type: 'UPDATE_USER',
+          attributes: body.user
+        })
+      })
+    }else{
+      // loginFail()
+      dispatcher.dispatch({
+        type: 'LOGIN_FAIL',
+      })
+    }
   })
 }
+
+// export function updateUser(attributes){
+//   dispatcher.dispatch({
+//     type: 'UPDATE_USER',
+//     attributes: attributes
+//   })
+// }
 
 export function logout(){
   dispatcher.dispatch({
@@ -41,15 +72,42 @@ export function logout(){
   })
 }
 
-export function registrationFail(errors){
-  dispatcher.dispatch({
-    type: 'REGISTRATION_FAIL',
-    errors: errors
-  })
-}
+// export function registrationFail(errors){
+//   dispatcher.dispatch({
+//     type: 'REGISTRATION_FAIL',
+//     errors: errors
+//   })
+// }
 
-export function loginFail(){
-  dispatcher.dispatch({
-    type: 'LOGIN_FAIL',
-  })
-}
+// export function loginFail(){
+//   dispatcher.dispatch({
+//     type: 'LOGIN_FAIL',
+//   })
+// }
+
+export function processRegistration(attributes){
+    const params = {
+      method: 'POST',
+      headers: {'Content-Type': 'application/json'},
+      body: JSON.stringify(attributes)
+    }
+    fetch(`${baseUrl}signup`, params).then((response)=>{
+      if(response.ok){
+        response.json().then((body)=>{
+          // updateUser(body.user)
+          dispatcher.dispatch({
+            type: 'UPDATE_USER',
+            attributes: body.user
+          })
+        })
+      }else{
+        response.json().then((body)=>{
+          // registrationFail(body.errors)
+          dispatcher.dispatch({
+            type: 'REGISTRATION_FAIL',
+            errors: body.errors
+          })
+        })
+      }
+    })
+  }
